# BUG/PERF: SeriesGroupBy.value_counts sorting bug and categorical performance